+++ src/components/index.ts
@@ -1,2 +1,2 @@
-import { Footer, Header, Login, Card, CardSection, Spinner, Profile } from "src/components/Layout/index"
-export { Footer, Header, Login, Card, CardSection, Spinner, Profile }
+import { Footer, Header, Login, Card, CardSection, Spinner, Profile, VideoLine } from "src/components/Layout/index"
+export { Footer, Header, Login, Card, CardSection, Spinner, Profile, VideoLine }
